refactor: encode_ext_lua_type
(1) If an object to-be-packed has a metatable and a __pack function, use
it instead of referencing the extended-type registry. Intent is to
simplify integration with custom userdata types.

(2) Oversight in userdata and lightuserdata sharing the same typename by
default; offer some additional level of granularity for future C-APIs

At this point, consider redesigning mp_get_type_extension and
mp_set_type_extension